+++ src/models/theatre.model.ts
