+++ src/routes/movies.route.ts
@@ -1,0 +1,8 @@
+import { Router } from "express";
+import moviesInCityController from "../controllers/movies.controller";
+
+const router = Router();
+
+router.route("/").get([moviesInCityController.getAllMoviesInCity]);
+
+export default router;
