+++ src/services/movies.service.ts
@@ -1,0 +1,16 @@
+import { Movie } from "../models/movie.model";
+import { MoviesInCityRepository } from "../repositories/movies.repository";
+
+export class MoviesInCityServices {
+  moviesInCityRepository: MoviesInCityRepository;
+
+  constructor() {
+    this.moviesInCityRepository = new MoviesInCityRepository();
+  }
+  async getAllMoviesInCity(id: string): Promise<Movie[]> {
+    return await this.moviesInCityRepository.getAllMoviesInCity(id);
+  }
+}
+
+const moviesInCityServices = new MoviesInCityServices();
+export default moviesInCityServices;
